+++ instrumentation/opentelemetry-instrumentation-aiohttp-client/tests/test_aiohttp_client_integration.py
@@ -15,6 +15,7 @@
 import asyncio
 import contextlib
 import typing
+import unittest
 import urllib.parse
 from http import HTTPStatus
 from unittest import mock
@@ -22,15 +23,39 @@ from unittest import mock
 import aiohttp
 import aiohttp.test_utils
 import yarl
+from pkg_resources import iter_entry_points
 
-import opentelemetry.instrumentation.aiohttp_client
+from opentelemetry import context
+from opentelemetry.instrumentation import aiohttp_client
+from opentelemetry.instrumentation.aiohttp_client import (
+    AioHttpClientInstrumentor,
+)
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace.status import StatusCanonicalCode
 
 
-class TestAioHttpIntegration(TestBase):
-    maxDiff = None
+def run_with_test_server(
+    runnable: typing.Callable, url: str, handler: typing.Callable
+) -> typing.Tuple[str, int]:
+    async def do_request():
+        app = aiohttp.web.Application()
+        parsed_url = urllib.parse.urlparse(url)
+        app.add_routes([aiohttp.web.get(parsed_url.path, handler)])
+        app.add_routes([aiohttp.web.post(parsed_url.path, handler)])
+        app.add_routes([aiohttp.web.patch(parsed_url.path, handler)])
 
+        with contextlib.suppress(aiohttp.ClientError):
+            async with aiohttp.test_utils.TestServer(app) as server:
+                netloc = (server.host, server.port)
+                await server.start_server()
+                await runnable(server)
+        return netloc
+
+    loop = asyncio.get_event_loop()
+    return loop.run_until_complete(do_request())
+
+
+class TestAioHttpIntegration(TestBase):
     def assert_spans(self, spans):
         self.assertEqual(
             [
@@ -54,9 +79,7 @@ class TestAioHttpIntegration(TestBase):
         ):
             with self.subTest(url=url):
                 params = aiohttp.TraceRequestStartParams("METHOD", url, {})
-                actual = opentelemetry.instrumentation.aiohttp_client.url_path_span_name(
-                    params
-                )
+                actual = aiohttp_client.url_path_span_name(params)
                 self.assertEqual(actual, expected)
                 self.assertIsInstance(actual, str)
 
@@ -71,33 +94,20 @@ class TestAioHttpIntegration(TestBase):
     ) -> typing.Tuple[str, int]:
         """Helper to start an aiohttp test server and send an actual HTTP request to it."""
 
-        async def do_request():
-            async def default_handler(request):
-                assert "traceparent" in request.headers
-                return aiohttp.web.Response(status=int(status_code))
+        async def default_handler(request):
+            assert "traceparent" in request.headers
+            return aiohttp.web.Response(status=int(status_code))
 
-            handler = request_handler or default_handler
+        async def client_request(server: aiohttp.test_utils.TestServer):
+            async with aiohttp.test_utils.TestClient(
+                server, trace_configs=[trace_config]
+            ) as client:
+                await client.request(
+                    method, url, trace_request_ctx={}, **kwargs
+                )
 
-            app = aiohttp.web.Application()
-            parsed_url = urllib.parse.urlparse(url)
-            app.add_routes([aiohttp.web.get(parsed_url.path, handler)])
-            app.add_routes([aiohttp.web.post(parsed_url.path, handler)])
-            app.add_routes([aiohttp.web.patch(parsed_url.path, handler)])
-
-            with contextlib.suppress(aiohttp.ClientError):
-                async with aiohttp.test_utils.TestServer(app) as server:
-                    netloc = (server.host, server.port)
-                    async with aiohttp.test_utils.TestClient(
-                        server, trace_configs=[trace_config]
-                    ) as client:
-                        await client.start_server()
-                        await client.request(
-                            method, url, trace_request_ctx={}, **kwargs
-                        )
-            return netloc
-
-        loop = asyncio.get_event_loop()
-        return loop.run_until_complete(do_request())
+        handler = request_handler or default_handler
+        return run_with_test_server(client_request, url, handler)
 
     def test_status_codes(self):
         for status_code, span_status in (
@@ -111,7 +121,7 @@ class TestAioHttpIntegration(TestBase):
         ):
             with self.subTest(status_code=status_code):
                 host, port = self._http_request(
-                    trace_config=opentelemetry.instrumentation.aiohttp_client.create_trace_config(),
+                    trace_config=aiohttp_client.create_trace_config(),
                     url="/test-path?query=param#foobar",
                     status_code=status_code,
                 )
@@ -144,7 +154,7 @@ class TestAioHttpIntegration(TestBase):
         with mock.patch("opentelemetry.trace.get_tracer"):
             # pylint: disable=W0612
             host, port = self._http_request(
-                trace_config=opentelemetry.instrumentation.aiohttp_client.create_trace_config(),
+                trace_config=aiohttp_client.create_trace_config(),
                 url="/test-path?query=param#foobar",
             )
             self.assertFalse(mock_span.is_recording())
@@ -166,7 +176,7 @@ class TestAioHttpIntegration(TestBase):
         ):
             with self.subTest(span_name=span_name, method=method, path=path):
                 host, port = self._http_request(
-                    trace_config=opentelemetry.instrumentation.aiohttp_client.create_trace_config(
+                    trace_config=aiohttp_client.create_trace_config(
                         span_name=span_name
                     ),
                     method=method,
@@ -199,7 +209,7 @@ class TestAioHttpIntegration(TestBase):
             return str(url.with_query(None))
 
         host, port = self._http_request(
-            trace_config=opentelemetry.instrumentation.aiohttp_client.create_trace_config(
+            trace_config=aiohttp_client.create_trace_config(
                 url_filter=strip_query_params
             ),
             url="/some/path?query=param&other=param2",
@@ -225,9 +235,7 @@ class TestAioHttpIntegration(TestBase):
         )
 
     def test_connection_errors(self):
-        trace_configs = [
-            opentelemetry.instrumentation.aiohttp_client.create_trace_config()
-        ]
+        trace_configs = [aiohttp_client.create_trace_config()]
 
         for url, expected_status in (
             ("http://this-is-unknown.local/", StatusCanonicalCode.UNKNOWN),
@@ -237,7 +245,7 @@ class TestAioHttpIntegration(TestBase):
 
                 async def do_request(url):
                     async with aiohttp.ClientSession(
-                        trace_configs=trace_configs
+                        trace_configs=trace_configs,
                     ) as session:
                         async with session.get(url):
                             pass
@@ -268,7 +276,7 @@ class TestAioHttpIntegration(TestBase):
             return aiohttp.web.Response()
 
         host, port = self._http_request(
-            trace_config=opentelemetry.instrumentation.aiohttp_client.create_trace_config(),
+            trace_config=aiohttp_client.create_trace_config(),
             url="/test_timeout",
             request_handler=request_handler,
             timeout=aiohttp.ClientTimeout(sock_read=0.01),
@@ -298,7 +306,7 @@ class TestAioHttpIntegration(TestBase):
             raise aiohttp.web.HTTPFound(location=location)
 
         host, port = self._http_request(
-            trace_config=opentelemetry.instrumentation.aiohttp_client.create_trace_config(),
+            trace_config=aiohttp_client.create_trace_config(),
             url="/test_too_many_redirects",
             request_handler=request_handler,
             max_redirects=2,
@@ -319,3 +327,177 @@ class TestAioHttpIntegration(TestBase):
                 )
             ]
         )
+
+
+class TestAioHttpClientInstrumentor(TestBase):
+    URL = "/test-path"
+
+    def setUp(self):
+        super().setUp()
+        AioHttpClientInstrumentor().instrument()
+
+    def tearDown(self):
+        super().tearDown()
+        AioHttpClientInstrumentor().uninstrument()
+
+    @staticmethod
+    # pylint:disable=unused-argument
+    async def default_handler(request):
+        return aiohttp.web.Response(status=int(200))
+
+    @staticmethod
+    def get_default_request(url: str = URL):
+        async def default_request(server: aiohttp.test_utils.TestServer):
+            async with aiohttp.test_utils.TestClient(server) as session:
+                await session.get(url)
+
+        return default_request
+
+    def assert_spans(self, num_spans: int):
+        finished_spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(num_spans, len(finished_spans))
+        if num_spans == 0:
+            return None
+        if num_spans == 1:
+            return finished_spans[0]
+        return finished_spans
+
+    def test_instrument(self):
+        host, port = run_with_test_server(
+            self.get_default_request(), self.URL, self.default_handler
+        )
+        span = self.assert_spans(1)
+        self.assertEqual("http", span.attributes["component"])
+        self.assertEqual("GET", span.attributes["http.method"])
+        self.assertEqual(
+            "http://{}:{}/test-path".format(host, port),
+            span.attributes["http.url"],
+        )
+        self.assertEqual(200, span.attributes["http.status_code"])
+        self.assertEqual("OK", span.attributes["http.status_text"])
+
+    def test_instrument_with_existing_trace_config(self):
+        trace_config = aiohttp.TraceConfig()
+
+        async def create_session(server: aiohttp.test_utils.TestServer):
+            async with aiohttp.test_utils.TestClient(
+                server, trace_configs=[trace_config]
+            ) as client:
+                # pylint:disable=protected-access
+                trace_configs = client.session._trace_configs
+                self.assertEqual(2, len(trace_configs))
+                self.assertTrue(trace_config in trace_configs)
+                async with client as session:
+                    await session.get(TestAioHttpClientInstrumentor.URL)
+
+        run_with_test_server(create_session, self.URL, self.default_handler)
+        self.assert_spans(1)
+
+    def test_uninstrument(self):
+        AioHttpClientInstrumentor().uninstrument()
+        run_with_test_server(
+            self.get_default_request(), self.URL, self.default_handler
+        )
+
+        self.assert_spans(0)
+
+        AioHttpClientInstrumentor().instrument()
+        run_with_test_server(
+            self.get_default_request(), self.URL, self.default_handler
+        )
+        self.assert_spans(1)
+
+    def test_uninstrument_session(self):
+        async def uninstrument_request(server: aiohttp.test_utils.TestServer):
+            client = aiohttp.test_utils.TestClient(server)
+            AioHttpClientInstrumentor().uninstrument_session(client.session)
+            async with client as session:
+                await session.get(self.URL)
+
+        run_with_test_server(
+            uninstrument_request, self.URL, self.default_handler
+        )
+        self.assert_spans(0)
+
+        run_with_test_server(
+            self.get_default_request(), self.URL, self.default_handler
+        )
+        self.assert_spans(1)
+
+    def test_suppress_instrumentation(self):
+        token = context.attach(
+            context.set_value("suppress_instrumentation", True)
+        )
+        try:
+            run_with_test_server(
+                self.get_default_request(), self.URL, self.default_handler
+            )
+        finally:
+            context.detach(token)
+        self.assert_spans(0)
+
+    @staticmethod
+    async def suppressed_request(server: aiohttp.test_utils.TestServer):
+        async with aiohttp.test_utils.TestClient(server) as client:
+            token = context.attach(
+                context.set_value("suppress_instrumentation", True)
+            )
+            await client.get(TestAioHttpClientInstrumentor.URL)
+            context.detach(token)
+
+    def test_suppress_instrumentation_after_creation(self):
+        run_with_test_server(
+            self.suppressed_request, self.URL, self.default_handler
+        )
+        self.assert_spans(0)
+
+    def test_suppress_instrumentation_with_server_exception(self):
+        # pylint:disable=unused-argument
+        async def raising_handler(request):
+            raise aiohttp.web.HTTPFound(location=self.URL)
+
+        run_with_test_server(
+            self.suppressed_request, self.URL, raising_handler
+        )
+        self.assert_spans(0)
+
+    def test_url_filter(self):
+        def strip_query_params(url: yarl.URL) -> str:
+            return str(url.with_query(None))
+
+        AioHttpClientInstrumentor().uninstrument()
+        AioHttpClientInstrumentor().instrument(url_filter=strip_query_params)
+
+        url = "/test-path?query=params"
+        host, port = run_with_test_server(
+            self.get_default_request(url), url, self.default_handler
+        )
+        span = self.assert_spans(1)
+        self.assertEqual(
+            "http://{}:{}/test-path".format(host, port),
+            span.attributes["http.url"],
+        )
+
+    def test_span_name(self):
+        def span_name_callback(params: aiohttp.TraceRequestStartParams) -> str:
+            return "{} - {}".format(params.method, params.url.path)
+
+        AioHttpClientInstrumentor().uninstrument()
+        AioHttpClientInstrumentor().instrument(span_name=span_name_callback)
+
+        url = "/test-path"
+        run_with_test_server(
+            self.get_default_request(url), url, self.default_handler
+        )
+        span = self.assert_spans(1)
+        self.assertEqual("GET - /test-path", span.name)
+
+
+class TestLoadingAioHttpInstrumentor(unittest.TestCase):
+    def test_loading_instrumentor(self):
+        entry_points = iter_entry_points(
+            "opentelemetry_instrumentor", "aiohttp-client"
+        )
+
+        instrumentor = next(entry_points).load()()
+        self.assertIsInstance(instrumentor, AioHttpClientInstrumentor)